Further improved seperation of concerns between Project.js and DashboardSection.js
