+++ viewer/src/components/Project.js
@@ -3,42 +3,9 @@ import Dashboard from "./Dashboard";
 import Loading from "./Loading";
 import { Container } from "./styled/Container";
 import { HiTemplate } from 'react-icons/hi';
-import DashboardSection from './dashboard/DashboardSection';
+import DashboardSection, { organizeDashboardsByLevel } from './dashboard/DashboardSection';
 import FilterBar from './dashboard/FilterBar';
 import { fetchDashboardThumbnails } from '../services/thumbnailService';
-
-const organizeDashboardsByLevel = (dashboards) => {
-  if (!dashboards?.length) return {};
-
-  // Initialize levels
-  const levels = {
-    L0: [],
-    L1: [],
-    L2: [],
-    L3: [],
-    L4: [],
-    unassigned: []
-  };
-
-  // Sort dashboards into levels
-  dashboards.forEach(dashboard => {
-    if (!dashboard.level) {
-      levels.unassigned.push(dashboard);
-    } else {
-      levels[dashboard.level].push(dashboard);
-    }
-  });
-
-  // Sort each level alphabetically
-  Object.keys(levels).forEach(level => {
-    levels[level].sort((a, b) => a.name.localeCompare(b.name));
-  });
-
-  // Filter out empty levels
-  return Object.fromEntries(
-    Object.entries(levels).filter(([_, dashboards]) => dashboards.length > 0)
-  );
-};
 
 function Project(props) {
   const [searchTerm, setSearchTerm] = useState('');
@@ -49,23 +16,6 @@ function Project(props) {
   useEffect(() => {
     window.scrollTo(0, 0);
   }, [props.dashboardName]);
-
-  const dashboardsByLevel = useMemo(() => 
-    organizeDashboardsByLevel(props.dashboards),
-    [props.dashboards]
-  );
-
-  // Load existing thumbnails from API
-  useEffect(() => {
-    if (Object.keys(dashboardsByLevel).length > 0) {
-      fetchDashboardThumbnails(dashboardsByLevel, (dashboardName, thumbnail) => {
-        setThumbnails(prev => ({
-          ...prev,
-          [dashboardName]: thumbnail
-        }));
-      });
-    }
-  }, [dashboardsByLevel]);
 
   const availableTags = useMemo(() => {
     if (!props.dashboards) return [];
@@ -88,6 +38,23 @@ function Project(props) {
       return matchesSearch && matchesTags;
     });
   }, [props.dashboards, searchTerm, selectedTags]);
+
+  const dashboardsByLevel = useMemo(() => 
+    organizeDashboardsByLevel(filteredDashboards),
+    [filteredDashboards]
+  );
+
+  // Load existing thumbnails from API
+  useEffect(() => {
+    if (Object.keys(dashboardsByLevel).length > 0) {
+      fetchDashboardThumbnails(dashboardsByLevel, (dashboardName, thumbnail) => {
+        setThumbnails(prev => ({
+          ...prev,
+          [dashboardName]: thumbnail
+        }));
+      });
+    }
+  }, [dashboardsByLevel]);
 
   const renderLoading = () => {
     return <Loading />;
